+++ tests/test_worker.ts
@@ -14,8 +14,8 @@ import {
   DelayedError,
   WaitingError,
 } from '../src/classes';
-import { KeepJobs, MinimalJob } from '../src/interfaces';
-import { JobsOptions } from '../src/types';
+import { MinimalJob } from '../src/interfaces';
+import { JobsOptions, KeepJobs } from '../src/types';
 import {
   delay,
   isRedisVersionLowerThan,
@@ -1281,6 +1281,7 @@ describe('workers', function () {
       fail?: boolean,
     ) {
       const clock = sinon.useFakeTimers({
+        shouldClearNativeTimers: true,
         toFake: timerMethodsForFaking as any,
       });
       clock.reset();
@@ -1359,6 +1360,7 @@ describe('workers', function () {
       fail?: boolean,
     ) {
       const clock = sinon.useFakeTimers({
+        shouldClearNativeTimers: true,
         toFake: timerMethodsForFaking as any,
       });
       clock.reset();
@@ -1382,37 +1384,42 @@ describe('workers', function () {
       const datas = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14];
 
       // eslint-disable-next-line prefer-const
-      let jobIds;
+      let jobIds: (string | undefined)[];
 
-      const processing = new Promise<void>(resolve => {
+      const processing = new Promise<void>((resolve, reject) => {
         worker.on(fail ? 'failed' : 'completed', async job => {
-          clock.tick(1000);
+          try {
+            clock.tick(1000);
 
-          if (job.data == 14) {
-            const counts = await queue.getJobCounts(
-              fail ? 'failed' : 'completed',
-            );
+            if (job.data == 14) {
+              const counts = await queue.getJobCounts(
+                fail ? 'failed' : 'completed',
+              );
 
-            if (fail) {
-              expect(counts.failed).to.be.equal(expectedCount);
-            } else {
-              expect(counts.completed).to.be.equal(expectedCount);
+              if (fail) {
+                expect(counts.failed).to.be.equal(expectedCount);
+              } else {
+                expect(counts.completed).to.be.equal(expectedCount);
+              }
+
+              await Promise.all(
+                jobIds.map(async (jobId, index) => {
+                  const job = await queue.getJob(jobId);
+                  const logs = await queue.getJobLogs(jobId);
+                  if (index >= datas.length - expectedCount) {
+                    expect(job).to.not.be.equal(undefined);
+                    expect(logs.logs).to.not.be.empty;
+                  } else {
+                    expect(job).to.be.equal(undefined);
+                    expect(logs.logs).to.be.empty;
+                  }
+                }),
+              );
+              resolve();
             }
-
-            await Promise.all(
-              jobIds.map(async (jobId, index) => {
-                const job = await queue.getJob(jobId);
-                const logs = await queue.getJobLogs(jobId);
-                if (index >= datas.length - expectedCount) {
-                  expect(job).to.not.be.equal(undefined);
-                  expect(logs.logs).to.not.be.empty;
-                } else {
-                  expect(job).to.be.equal(undefined);
-                  expect(logs.logs).to.be.empty;
-                }
-              }),
-            );
-            resolve();
+          } catch (error) {
+            console.log(error);
+            reject(error);
           }
         });
       });
@@ -1693,6 +1700,12 @@ describe('workers', function () {
         const age = 7;
         const count = 5;
         await testWorkerRemoveOnFinish({ age, count }, count);
+      });
+
+      it('should keep jobs with age 10 and limit 2 with removeOnComplete', async () => {
+        const age = 10;
+        const limit = 2;
+        await testWorkerRemoveOnFinish({ age, limit }, age);
       });
 
       it('should keep of jobs newer than specified and up to a count fail with removeOnFail', async () => {